Bugfix
Fixed error where if the start date was within a week of the end of the month it would attempt to set a date beyond the end of the month without rolling over to the next one. Historically has not been an issue as the semester / trimester usually started at the beginning of March, however this year begins at the end of February and was attempting to roll over.